FEAT: Created chimeara process object
To be able to re-establish contact between server and its daemons, the
server needs a basic interface similar to the extended process interface
used in the RPC_Job. Thus common features of true and fake were moved to
the fake class. The RPC_Job was set to inherit both
multiprocessing.Process and the Fake class. Since initiation needed
different codes and interfaces, duplicate instructions were made for
init.