FEAT: Made pipe effectors handle broken pipes
Getting send and recieve exceptions changes effector state to broken.
If send messages on child side requires parent, child will check if
server is online and if not spawn a new process.
When server is online, it will request a new pipe from the rpc-side of
the server. The server will set up a fake internal job-process that
will mimic the original job and pipe will be sent to child.
This only occurs if the child can identify itself as a job that the
server knows.