:star: Hidden 16bit support
Importing 16bit images will currently deprecate the higher resolution
but is supported. For future expanding of functionality to 16bit
it should only require to remove the `dtype=np.uint8` of the
`load_image_to_numpy` function.

Of course it is also needed to give user option to record 16bit
images when starting experiments and extending under the hood
functionality for getting this info to the scanners