Remove docker-compose.override.yml
This is a proposal to make working with Scan-o-Matic easier.
The idea is to only include a compose file that is minimal and "safe"
(in the sense of not creating weird directory in your root or bind on
ports that may not be available) and let user/developers use the
override file for local modification whithout bothering git about it
(for instance, I like mounting the js files from my work directory when
working on the UI.)

Removing the fixed port binding might be annoying if you restart the
system often, but it isn't possible to remove a port binding from the
override file, which is annoying as well. (It is possible to add a port
binding on the other hand).

One may also use the override file to create weird directories on one's
machine if that's a thing one cares for.