Added test of generic name
Random accepts any hasahble object as seed. This results in very large
seeds and for some reason seeding with large numbers gives
inconsistent results in tox and/or pytest. This might be fixable by
setting pytest's seed explicitly as well in tox.ini, but for now I'm
happy with having tested it in the terminal, where it does work
consistently.