Split travis build in two stages
This uses a beta feature on travis that allows splitting a build in
successive stages. (https://docs.travis-ci.com/user/build-stages/)

The build is split in two stages: first the (white box) tests, then
docker build and system tests.

Moreover, the tests are divided in two jobs: one running the python
tests and one running the javascript tests.

Given the overhead of setting up jobs on travis, dividing the build
doesn't really give us any speed boost but it makes it much nicer to
look at the result and allow restarting a single job.

On the other hand, the last thing I did was to remove the extra option
to `docker-compose build` and in the last two builds it went twice as
fast.  That's not a lot of data points but it'd be cool if it were to be
true :pray:

Some other relevant changes:
- this disables coveralls, since it isn't capable of merging coverage
  reports from the different jobs (js and python tests).
- to make tox a bit faster I put the system tests in a separate test
  environment in tox.  That reduces the number of unnecessary
  dependencies that are going to be installed.  I'm not sure it's a huge
  gain but it also makes it easier to run only unit/integration tests
  locally (that's `tox -e py27`, simply running `tox` will still run
  everything)